[#1599] Remove OT::Utils::SecureNumbers
In favour of Familia's where they originated. We are moving over
generate_short_id there as well.
